# Fix get_field() performance bottleneck on local fields (~40%)