Hello,

I was working on a faster alternative to `get_field()` function when I discovered a performance bottleneck on local fields.

Description:

The problem come from `acf_get_field()` function, which is called every time `get_field()` is invoked. What it does for each fields:

- Check first if the field is local
- If the local field is found:
- - Apply `acf_validate_field()`
- - Apply `$field['prefix'] = 'acf';`
- - Apply `acf/load_field` filters
- - Push the field into the buffering `store('fields')` for the next call.

I was able to gain ~40% performance on `get_field()` calls with local fields by doing so:

If the field is local: Do not process `validate`, `prefix`, `acf/load_field` in `acf_get_field()`.
Inject it during the local fields registration, in `acf_add_local_fields()`:

- Apply `acf_validate_field()`
- Apply `$field['prefix'] = 'acf';`
- Apply `acf/load_field` filters

Finally, in `acf_get_field()`, if the field is local, then directly return the local field.

I ran a benchmark on local environment, on a page with 200 `get_field()` on 200 different fields (local PHP field group):
- 0.0272 sec
- 0.0308 sec
- 0.0260 sec

With the fix in this pull request:
- 0.0179 sec
- 0.0185 sec
- 0.0167 sec

I ran some tests on `acf/load_field` filter, and it works as expected.

Hope it helps :)

Regards.