@elliotcondon 2019-12-02
Hi @acf-extended 

Thanks for your contribution.
Elliot here - ACF dev.

The 40% performance boost here is impressive, but also quite confusing. I suspect this boost might be a "false positive" gained by reducing functionality/compatibility. Let me explain.

The "acf/load_field" filter is an extremely important one. It allows complex field types such as the Repeater and Flexible Content to load their sub field, amongst other functionality. It is only ever run once per field, when that field is first loaded, even if that field is "loaded" multiple times.

It is important that this filter is run after the ACF field types have been registered, otherwise, the field type (Repeater / Flex) will not receive their chance to modify the `$field` before it is placed in temp-cache storage.

I suspect that your PR is unintentionally preventing these classes from modifying the `$field` variable during "acf/load_field" due to the change in timing. I may be wrong here - as I have not yet tested this code.

If my suspicions are correct, this PR could cause many PHP errors due to incomplete `$field` properties and missing sub field data.

Logically, it does not seem correct for a "performance boost" to be achieved by running extra functionality on page load. This PR forces the "acf/load_field" filter to be run on every registered field, whether that field is actually loaded on the page. The current implementation would not doe this, which leads me to think this is a "false positive" type situation.

Can you please perform some more testing with this PR?
- Is there a difference in performance between registering fields in root `functions.php` file vs "acf/include_fields" action?
- Does this PR cause individually added fields (`acf_add_local_field()`) to be lost when the parent field is rendered/loaded?

Thanks mate. Interested in your results :)