@acf-extended 2019-12-02
Hello,

Thanks for taking the time to write some explanations. I'll run extensive tests on complex fields and test the cases you described.

Meanwhile, I rename the PR, and will get back ASAP.

Regards.